+++ plugins/modules/sentinelone_sites.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
 
-# Copyright: (c) 2023, Marco Wester <marco.wester@sva.de>
+# Copyright: (c) 2024, Marco Wester <marco.wester@sva.de>
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
@@ -94,7 +94,7 @@ notes:
 EXAMPLES = r'''
 ---
 - name: Create / update site
-  sentinelone_sites:
+  sva.sentinelone.sentinelone_sites:
     console_url: "https://XXXXX.sentinelone.net"
     token: "XXXXXXXXXXXXXXXXXXXXXXXXXXX"
     name: "test"
@@ -102,7 +102,7 @@ EXAMPLES = r'''
     expiration_date: "2022-06-01T12:00+01:00"
     description: "Testsite"
 - name: Delete site
-  sentinelone_sites:
+  sva.sentinelone.sentinelone_sites:
     state: "absent"
     console_url: "https://XXXXX.sentinelone.net"
     token: "XXXXXXXXXXXXXXXXXXXXXXXXXXX"
@@ -373,7 +373,7 @@ def run_module():
 
     result = dict(
         changed=False,
-        original_message=str(diffs),
+        original_message=diffs,
         message=basic_message
     )